refactor: implement repository pattern for modularity and code reuse
- Moved business logic from service classes (`AuthService`, `ProductCategoryService`, `ShippingMethodService`, `ShippingAddressService`) to separate repository implementations.
- Added new repository implementation classes: `AuthRepositoryImpl`, `ProductCategoryRepositoryImpl`, `ShopCategoryRepositoryImpl`, `BrandRepositoryImpl`, and `ShippingMethodRepositoryImpl`.
- Updated service classes to delegate calls to respective repository interfaces for better modularity and separation of concerns.
- Removed redundant and unused service methods.
- Improved maintainability and testability of the codebase.
