+++ src/main/kotlin/com/piashcse/feature/consent/ConsentService.kt
@@ -1,89 +1,3 @@
 package com.piashcse.feature.consent
 
-import com.piashcse.constants.Message
-import com.piashcse.constants.PolicyType
-import com.piashcse.database.entities.*
-import com.piashcse.model.request.PolicyConsentRequest
-import com.piashcse.model.response.UserPolicyConsentResponse
-import com.piashcse.utils.extension.*
-import com.piashcse.utils.validator.ValidationException
-import org.jetbrains.exposed.v1.core.and
-import org.jetbrains.exposed.v1.core.eq
-import java.time.LocalDateTime
-
-class ConsentService : ConsentRepository {
-    /**
-     * Records user consent to a policy
-     */
-    override suspend fun recordConsent(
-        userId: String,
-        consentRequest: PolicyConsentRequest,
-    ): UserPolicyConsentResponse =
-        query {
-            userId.requireNotBlank("User ID")
-            consentRequest.policyId.requireNotBlank("Policy ID")
-
-            // Verify user and policy exist
-            val user = UserDAO.findById(userId) ?: userId.throwNotFound("User")
-            val policy =
-                PolicyDocumentDAO.findById(consentRequest.policyId)
-                    ?: consentRequest.policyId.throwNotFound("Policy")
-
-            // Check if consent already exists, if so update it
-            val existingConsent =
-                PolicyConsentDAO.find {
-                    PolicyConsentTable.userId eq user.id and (PolicyConsentTable.policyId eq policy.id)
-                }.firstOrNull()
-
-            val consent =
-                existingConsent?.apply {
-                    // Update existing consent
-                    consentDate = LocalDateTime.now()
-                    ipAddress = consentRequest.ipAddress
-                    userAgent = consentRequest.userAgent
-                } ?: PolicyConsentDAO.new {
-                    // Create new consent
-                    this.userId = user.id
-                    policyId = policy
-                    consentDate = LocalDateTime.now()
-                    ipAddress = consentRequest.ipAddress
-                    userAgent = consentRequest.userAgent
-                }
-
-            // Return the response
-            consent.response()
-        }
-
-    /**
-     * Gets all consents for a user
-     */
-    override suspend fun getUserConsents(userId: String): List<UserPolicyConsentResponse> =
-        query {
-            val user = UserDAO.findById(userId) ?: userId.throwNotFound("User")
-
-            PolicyConsentDAO.find { PolicyConsentTable.userId eq user.id }
-                .map { it.response() }
-        }
-
-    /**
-     * Checks if a user has consented to a specific policy
-     */
-    override suspend fun hasUserConsented(
-        userId: String,
-        policyType: PolicyType,
-    ): Boolean =
-        query {
-            val user = UserDAO.findById(userId) ?: userId.throwNotFound("User")
-
-            // Find the active policy of the specified type
-            val activePolicy =
-                PolicyDocumentDAO.find {
-                    PolicyDocumentTable.type eq policyType and (PolicyDocumentTable.isActive eq true)
-                }.firstOrNull() ?: return@query false
-
-            // Check if user has consented to this policy
-            PolicyConsentDAO.find {
-                PolicyConsentTable.userId eq user.id and (PolicyConsentTable.policyId eq activePolicy.id)
-            }.firstOrNull() != null
-        }
-}
+class ConsentService(private val repo: ConsentRepository) : ConsentRepository by repo
