+++ src/main/kotlin/com/piashcse/feature/product_category/ProductCategoryService.kt
@@ -1,89 +1,3 @@
 package com.piashcse.feature.product_category
 
-import com.piashcse.database.entities.ProductCategoryDAO
-import com.piashcse.database.entities.ProductCategoryTable
-import com.piashcse.model.response.ProductCategoryResponse
-import com.piashcse.utils.common.PaginatedResponse
-import com.piashcse.utils.extension.query
-import com.piashcse.utils.extension.throwConflict
-import com.piashcse.utils.extension.throwNotFound
-import com.piashcse.utils.extension.toPaginatedResponse
-import org.jetbrains.exposed.v1.core.eq
-import org.jetbrains.exposed.v1.jdbc.selectAll
-
-/**
- * Service for managing product categories.
- */
-class ProductCategoryService : ProductCategoryRepository {
-    /**
-     * Creates a new product category with the given category name.
-     *
-     * @param name The name of the category to be created.
-     * @return The created product category entity.
-     * @throws Exception if a category with the provided name already exists.
-     */
-    override suspend fun createCategory(name: String): ProductCategoryResponse =
-        query {
-            val isCategoryExist =
-                ProductCategoryDAO.find { ProductCategoryTable.name eq name }.firstOrNull()
-            isCategoryExist?.let {
-                throw name.throwConflict("Category")
-            } ?: ProductCategoryDAO.new {
-                this.name = name
-            }.response()
-        }
-
-    /**
-     * Retrieves the list of product categories with a limit on the number of categories returned.
-     *
-     * @param limit The maximum number of categories to retrieve.
-     * @return A list of product category entities.
-     */
-    override suspend fun getCategories(
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ProductCategoryResponse> =
-        query {
-            ProductCategoryTable.selectAll().toPaginatedResponse(limit, offset) {
-                ProductCategoryDAO.wrapRow(it).response()
-            }
-        }
-
-    /**
-     * Updates the name of an existing product category.
-     *
-     * @param categoryId The ID of the category to update.
-     * @param categoryName The new name for the category.
-     * @return The updated product category entity.
-     * @throws Exception if no category is found with the provided category ID.
-     */
-    override suspend fun updateCategory(
-        categoryId: String,
-        name: String,
-    ): ProductCategoryResponse =
-        query {
-            val isCategoryExist =
-                ProductCategoryDAO.findById(categoryId)
-            isCategoryExist?.let {
-                it.name = name
-                it.response()
-            } ?: categoryId.throwNotFound("Category")
-        }
-
-    /**
-     * Deletes an existing product category by its ID.
-     *
-     * @param categoryId The ID of the category to delete.
-     * @return The ID of the deleted category.
-     * @throws Exception if no category is found with the provided category ID.
-     */
-    override suspend fun deleteCategory(categoryId: String): String =
-        query {
-            val isCategoryExist =
-                ProductCategoryDAO.findById(categoryId)
-            isCategoryExist?.let {
-                isCategoryExist.delete()
-                categoryId
-            } ?: categoryId.throwNotFound("Category")
-        }
-}
+class ProductCategoryService(private val repo: ProductCategoryRepository) : ProductCategoryRepository by repo
