+++ src/main/kotlin/com/piashcse/feature/refund_request/RefundRequestRepository.kt
@@ -1,0 +1,42 @@
+package com.piashcse.feature.refund_request
+
+import com.piashcse.constants.UserType
+import com.piashcse.model.request.RefundRequestRequest
+import com.piashcse.model.request.ShipRefundRequest
+import com.piashcse.model.request.UpdateRefundStatusRequest
+import com.piashcse.model.response.RefundRequestResponse
+import com.piashcse.utils.common.PaginatedResponse
+
+interface RefundRequestRepository {
+    suspend fun createRefundRequest(
+        userId: String,
+        orderId: String,
+        request: RefundRequestRequest,
+    ): RefundRequestResponse
+
+    suspend fun getRefundsByOrderId(
+        orderId: String,
+        userId: String,
+        userType: UserType,
+        limit: Int = 20,
+        offset: Int = 0,
+    ): PaginatedResponse<RefundRequestResponse>
+
+    suspend fun getRefundById(
+        refundId: String,
+        userId: String,
+        userType: UserType,
+    ): RefundRequestResponse?
+
+    suspend fun updateRefundStatus(
+        refundId: String,
+        request: UpdateRefundStatusRequest,
+        userId: String,
+    ): RefundRequestResponse
+
+    suspend fun shipRefund(
+        refundId: String,
+        request: ShipRefundRequest,
+        userId: String,
+    ): RefundRequestResponse
+}
