+++ src/main/kotlin/com/piashcse/feature/shop/ShopService.kt
@@ -1,315 +1,3 @@
 package com.piashcse.feature.shop
 
-import com.piashcse.constants.Message
-import com.piashcse.constants.ShopStatus
-import com.piashcse.database.entities.*
-import com.piashcse.model.request.ShopRequest
-import com.piashcse.model.request.UpdateShopRequest
-import com.piashcse.model.response.ShopResponse
-import com.piashcse.utils.common.PaginatedResponse
-import com.piashcse.utils.extension.*
-import com.piashcse.utils.validator.ConflictException
-import com.piashcse.utils.validator.ForbiddenException
-import com.piashcse.utils.validator.NotFoundException
-import org.jetbrains.exposed.v1.core.SortOrder
-import org.jetbrains.exposed.v1.core.and
-import org.jetbrains.exposed.v1.core.dao.id.EntityID
-import org.jetbrains.exposed.v1.core.eq
-import org.jetbrains.exposed.v1.core.neq
-import org.jetbrains.exposed.v1.jdbc.andWhere
-import org.jetbrains.exposed.v1.jdbc.selectAll
-
-class ShopService : ShopRepository {
-    /**
-     * Creates a new shop for a seller.
-     *
-     * @param userId The user ID creating the shop.
-     * @param shopRequest The shop details to create.
-     * @return The created shop.
-     */
-    override suspend fun createShop(
-        userId: String,
-        shopRequest: ShopRequest,
-    ): ShopResponse =
-        query {
-            val user =
-                UserDAO.findById(userId)
-                    ?: userId.throwNotFound("User")
-
-            // Verify that the user is a seller by checking for a corresponding seller record
-            val seller = findSellerByUserId(userId)
-                ?: throw NotFoundException(Message.Errors.SELLER_REQUIRED)
-
-            val existingShop =
-                ShopDAO.find {
-                    ShopTable.userId eq userId
-                }.firstOrNull()
-
-            if (existingShop != null) {
-                // If user already has a shop, throw an error or update the existing one
-                // For this implementation, we'll throw an exception
-                throw ConflictException(Message.Shops.ALREADY_EXISTS)
-            }
-
-            val shop =
-                ShopDAO.new {
-                    this.userId = EntityID(userId, ShopTable)
-                    categoryId = EntityID(shopRequest.categoryId, ShopCategoryTable)
-                    name = shopRequest.name
-                    description = shopRequest.description
-                    address = shopRequest.address
-                    phone = shopRequest.phone
-                    email = shopRequest.email
-                    logo = shopRequest.logo
-                    coverImage = shopRequest.coverImage
-                    status = ShopStatus.PENDING // Initially pending approval
-                }
-
-            // Update the seller record to link to the new shop
-            seller.shopId = shop.id
-
-            shop.shopResponse()
-        }
-
-    /**
-     * Updates an existing shop.
-     *
-     * @param userId The user ID updating the shop.
-     * @param shopId The shop ID to update.
-     * @param shopRequest The shop details to update.
-     * @return The updated shop.
-     */
-    override suspend fun updateShop(
-        userId: String,
-        shopId: String,
-        shopRequest: UpdateShopRequest,
-    ): ShopResponse =
-        query {
-            val shop =
-                ShopDAO.findById(shopId)
-                    ?: shopId.throwNotFound("Shop")
-
-            shop.verifyOwnership(userId, "shop") { it.userId.value }
-
-            shop.apply {
-                name = shopRequest.name ?: name
-                description = shopRequest.description ?: description
-                address = shopRequest.address ?: address
-                phone = shopRequest.phone ?: phone
-                email = shopRequest.email ?: email
-                logo = shopRequest.logo ?: logo
-                coverImage = shopRequest.coverImage ?: coverImage
-            }
-
-            shop.shopResponse()
-        }
-
-    /**
-     * Gets a shop by ID.
-     *
-     * @param shopId The shop ID to retrieve.
-     * @return The shop details.
-     */
-    override suspend fun getShopById(shopId: String): ShopResponse? =
-        query {
-            val shop = ShopDAO.findById(shopId)
-            shop?.shopResponse()
-        }
-
-    /**
-     * Gets shops by user ID (seller).
-     *
-     * @param userId The user ID to retrieve shops for.
-     * @return A list of shops for the user.
-     */
-    override suspend fun getShopsByUser(
-        userId: String,
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ShopResponse> =
-        query {
-            ShopTable.selectAll().andWhere { ShopTable.userId eq userId }
-                .toPaginatedResponse(limit, offset) {
-                    ShopDAO.wrapRow(it).shopResponse()
-                }
-        }
-
-    /**
-     * Gets all shops with optional filtering.
-     *
-     * @param status Optional status to filter by.
-     * @param category Optional category to filter by.
-     * @param limit Number of shops to return.
-     * @return A list of shops matching the criteria.
-     */
-    override suspend fun getShops(
-        status: String?,
-        category: String?,
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ShopResponse> =
-        query {
-            // Validate status parameter
-            val statusEnum =
-                if (status != null) {
-                    try {
-                        ShopStatus.valueOf(status.uppercase())
-                    } catch (e: IllegalArgumentException) {
-                        throw NotFoundException(Message.Shops.invalidStatus(status))
-                    }
-                } else {
-                    null
-                }
-
-            val query = ShopTable.selectAll()
-
-            // Exclude rejected and suspended shops
-            query.andWhere { ShopTable.status neq ShopStatus.REJECTED }
-            query.andWhere { ShopTable.status neq ShopStatus.SUSPENDED }
-
-            if (statusEnum != null) {
-                query.andWhere { ShopTable.status eq statusEnum }
-            }
-            if (category != null) {
-                query.andWhere { ShopTable.categoryId eq EntityID(category, ShopCategoryTable) }
-            }
-
-            query.orderBy(ShopTable.createdAt to SortOrder.DESC).toPaginatedResponse(limit, offset) {
-                ShopDAO.wrapRow(it).shopResponse()
-            }
-        }
-
-    /**
-     * Gets shops by category.
-     *
-     * @param categoryId The category ID to filter by.
-     * @return A list of shops in the category.
-     */
-    override suspend fun getShopsByCategory(
-        categoryId: String,
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ShopResponse> =
-        query {
-            ShopTable.selectAll().andWhere {
-                ShopTable.categoryId eq categoryId and (ShopTable.status neq ShopStatus.REJECTED) and (ShopTable.status neq ShopStatus.SUSPENDED)
-            }.toPaginatedResponse(limit, offset) {
-                ShopDAO.wrapRow(it).shopResponse()
-            }
-        }
-
-    /**
-     * Gets featured shops.
-     *
-     * @return A list of featured shops.
-     */
-    override suspend fun getFeaturedShops(
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ShopResponse> =
-        query {
-            // For now, just return shops with highest ratings, in the future this could be more complex
-            ShopTable.selectAll().andWhere { ShopTable.status eq ShopStatus.APPROVED }
-                .orderBy(ShopTable.rating to SortOrder.DESC)
-                .toPaginatedResponse(limit, offset) {
-                    ShopDAO.wrapRow(it).shopResponse()
-                }
-        }
-
-    /**
-     * Gets shops by status.
-     *
-     * @param status The status to filter by.
-     * @return A list of shops with the specified status.
-     */
-    override suspend fun getShopsByStatus(
-        status: ShopStatus,
-        limit: Int,
-        offset: Int,
-    ): PaginatedResponse<ShopResponse> =
-        query {
-            ShopTable.selectAll().andWhere { ShopTable.status eq status }
-                .toPaginatedResponse(limit, offset) {
-                    ShopDAO.wrapRow(it).shopResponse()
-                }
-        }
-
-    /**
-     * Approves a shop application.
-     *
-     * @param shopId The shop ID to approve.
-     * @return The updated shop.
-     */
-    override suspend fun approveShop(shopId: String): ShopResponse =
-        query {
-            val shop =
-                ShopDAO.findById(shopId)
-                    ?: shopId.throwNotFound("ShopResponse")
-
-            shop.apply {
-                status = ShopStatus.APPROVED
-                // Note: createdAt should remain unchanged, only updatedAt will be changed automatically by the base class
-            }
-
-            shop.shopResponse()
-        }
-
-    /**
-     * Rejects a shop application.
-     *
-     * @param shopId The shop ID to reject.
-     * @return The updated shop.
-     */
-    override suspend fun rejectShop(shopId: String): ShopResponse =
-        query {
-            val shop =
-                ShopDAO.findById(shopId)
-                    ?: shopId.throwNotFound("ShopResponse")
-
-            shop.apply {
-                status = ShopStatus.REJECTED
-            }
-
-            shop.shopResponse()
-        }
-
-    /**
-     * Suspends a shop.
-     *
-     * @param shopId The shop ID to suspend.
-     * @return The updated shop.
-     */
-    override suspend fun suspendShop(shopId: String): ShopResponse =
-        query {
-            val shop =
-                ShopDAO.findById(shopId)
-                    ?: shopId.throwNotFound("ShopResponse")
-
-            shop.apply {
-                status = ShopStatus.SUSPENDED
-            }
-
-            shop.shopResponse()
-        }
-
-    /**
-     * Activates a suspended shop.
-     *
-     * @param shopId The shop ID to activate.
-     * @return The updated shop.
-     */
-    override suspend fun activateShop(shopId: String): ShopResponse =
-        query {
-            val shop =
-                ShopDAO.findById(shopId)
-                    ?: shopId.throwNotFound("ShopResponse")
-
-            shop.apply {
-                if (shop.status == ShopStatus.SUSPENDED) {
-                    status = ShopStatus.APPROVED
-                }
-            }
-
-            shop.shopResponse()
-        }
-}
+class ShopService(private val repo: ShopRepository) : ShopRepository by repo
